chore: update headerbar to slightly newer version
