+++ package.json
@@ -13,7 +13,7 @@
     "localize": "npm run extract-pot && d2-i18n-generate -n messaging-app -p ./i18n/ -o ./src/locales/"
   },
   "dependencies": {
-    "@dhis2/d2-ui-header-bar": "^1.1.1",
+    "@dhis2/ui": "^1.0.0-beta.14",
     "d2": "31.0.2",
     "d2-i18n": "^1.0.3",
     "d2-ui": "^29.0.25",
@@ -32,8 +32,7 @@
     "react-scripts": "3.0.1",
     "recompose": "^0.30.0",
     "redux": "^3.0.0",
-    "redux-thunk": "^2.3.0",
-    "rxjs": "^5.0.0"
+    "redux-thunk": "^2.3.0"
   },
   "devDependencies": {
     "@dhis2/code-style": "^1.3.0",
