+++ src/components/MessagingCenter/MessagingCenter.css
@@ -1,7 +1,6 @@
 .messaging-center {
     display: flex;
     flex-direction: column;
-    margin-top: 48px;
     width: 100%;
     height: calc(100vh - 55px);
     position: fixed;
